backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
dj-database-url==2.2.0
requests==2.32.3
cryptography==43.0.0
google-auth==2.34.0

# Auth
django-allauth==0.63.6
djangorestframework-simplejwt==5.3.1

# Realtime
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Async
celery==5.4.0
redis==5.0.8
